requirements.txt
# Weather data collection dependencies
requests>=2.28.0
pandas>=1.5.0
schedule>=1.2.0
python-dotenv>=0.19.0

# Database
# sqlite3 is included with Python standard library

# Optional: for enhanced data analysis
matplotlib>=3.6.0
numpy>=1.24.0